public IEnumerator Update(TankState tank);

+++ Assets/Script/Interfase/IHP.cs
@@ -2,7 +2,7 @@
 {
     public interface IHP
     {
-        public int HP { get; set; }
-        public int MaxHP { get; set; }
+        public short HP { get; set; }
+        public short MaxHP { get; set; }
     }
 }
